+++ src/hooks/useUrlQuery.ts
@@ -5,9 +5,15 @@ export default function useUrlQuery() {
 
   const page = searchParams.get('page');
   const orderby = searchParams.get('orderby');
+  const popular = searchParams.get('popular');
+  const variant = searchParams.get('variant');
+  const query = searchParams.get('query');
 
   return {
     page: page ? parseInt(page) : 0,
     orderby: orderby ? orderby : 'false',
+    popular: popular ? popular : 'false',
+    variant: variant ? variant : 'row',
+    query: query ? query : undefined,
   };
 }
